requirements.txt
mss
discord-webhook
python-dotenv
google-api-python-client
google-auth
